Extract document tooltip from document link widget to a separate view.
So it's possible to load the document tooltip via AJAX.
