+++ opengever/document/widgets/document_link.py
@@ -1,7 +1,5 @@
-from opengever.base.pdfconverter import is_pdfconverter_enabled
-from opengever.document.browser.download import DownloadConfirmationHelper
+from plone import api
 from plone.app.contentlisting.interfaces import IContentListingObject
-from plone.protect.utils import addTokenToUrl
 from zope.browserpage.viewpagetemplatefile import ViewPageTemplateFile
 from zope.globalrequest import getRequest
 
@@ -18,6 +16,9 @@ class DocumentLinkWidget(object):
     def get_url(self):
         return self.document.getURL()
 
+    def portal_url(self):
+        return api.portal.get().absolute_url()
+
     def get_css_class(self):
         classes = ['tabbedview-tooltip', 'document_link',
                    self.document.ContentTypeClass()]
@@ -29,26 +30,6 @@ class DocumentLinkWidget(object):
 
     def get_title(self):
         return self.document.Title().decode('utf-8')
-
-    def preview_link_available(self):
-        return self.document.is_document and is_pdfconverter_enabled()
-
-    def edit_metadata_link_available(self):
-        return not self.document.is_trashed
-
-    def checkout_and_edit_link_available(self):
-        return self.document.is_document
-
-    def checkout_and_edit_link(self):
-        return addTokenToUrl('{}/editing_document'.format(self.get_url()))
-
-    def download_link_available(self):
-        return self.document.is_document
-
-    def download_link(self):
-        dc_helper = DownloadConfirmationHelper()
-        return dc_helper.get_html_tag(self.get_url(),
-                                      additional_classes=['action-download'])
 
     def render(self):
         return self.template(self, self.request)
