+++ opengever/document/widgets/tooltip.pt
@@ -1,0 +1,40 @@
+<div class='tooltip-header' tal:content="context/title" />
+<ul class='tooltip-breadcrumb' tal:define="breadcrumbs view/document/get_breadcrumbs">
+  <li tal:repeat="breadcrumb breadcrumbs" tal:content="breadcrumb/Title" />
+</ul>
+
+<div class='tooltip-content'>
+  <ul class='tooltip-links'>
+    <li tal:condition="view/preview_link_available">
+      <a class="action-pdf"
+         tal:attributes="href string:${view/get_url}/@@download_pdfpreview"
+         i18n:translate="button_pdf_preview">
+        PDF Preview
+      </a>
+    </li>
+    <li tal:condition="view/edit_metadata_link_available">
+      <a class="action-edit"
+         tal:attributes="href string:${view/get_url}/edit_checker"
+         i18n:domain="plone" i18n:translate="">
+        Edit metadata
+      </a>
+    </li>
+    <li tal:condition="view/checkout_and_edit_link_available">
+      <a class="action-checkout"
+         tal:attributes="href view/checkout_and_edit_link"
+         i18n:domain="plone" i18n:translate="">
+        Checkout and edit
+      </a>
+    </li>
+    <li tal:condition="view/download_link_available">
+      <a class="action-download"
+         tal:replace="structure view/download_link" />
+    </li>
+  </ul>
+  <div class="preview" tal:condition="view/document/is_bumblebeeable">
+    <img tal:attributes="src view/document/get_preview_image_url; alt context/title"
+         width="200" class="file-preview" />
+  </div>
+</div>
+
+<div class='bottomImage'></div>
